+++ system/src/apps/CompendiumItemSelectors/_module.mjs
@@ -1,3 +1,4 @@
+export {default as AncestrySelector} from "./AncestrySelector.mjs";
 export {default as ArmorPropertySelector} from "./ArmorPropertySelector.mjs";
 export {default as LanguageSelector} from "./LanguageSelector.mjs";
 export {default as WeaponPropertySelector} from "./WeaponPropertySelector.mjs";