+++ system/src/migrations/updates/_module.mjs
@@ -11,3 +11,4 @@ export {default as Update_230911_1} from "./Update_230911_1.mjs";
 export {default as Update_230914_1} from "./Update_230914_1.mjs";
 export {default as Update_230915_1} from "./Update_230915_1.mjs";
 export {default as Update_230915_2} from "./Update_230915_2.mjs";
+export {default as Update_230916_1} from "./Update_230916_1.mjs";