+++ system/src/utils/UtilitySD.mjs
@@ -1,0 +1,30 @@
+export default class UtilitySD {
+	/**
+	 * Creates de-duplicated lists of Selected and Unselected Items.
+	 *
+	 * @param {allItems} Array A list of all available skills
+	 * @param {items} Array A list of currently selected skills
+	 *
+	 * @returns {Promise} Promise which represents an array containing both the
+	 * selected and unselected skill arrays
+	 */
+	static async getDedupedSelectedItems(allItems, items) {
+		const unselectedItems = [];
+		const selectedItems = [];
+
+		allItems.forEach(item => {
+			if (!items.includes(item.uuid)) {
+				unselectedItems.push(item);
+			}
+		});
+
+		for (const itemUuid of items) {
+			selectedItems.push(await fromUuid(itemUuid));
+		}
+
+		selectedItems.sort((a, b) => a.name.localeCompare(b.name));
+
+		return [selectedItems, unselectedItems];
+	}
+
+}